feat: remove the internal package and move to pkg structure
