+++ cmd/pkg/cmdname_krew.go
@@ -1,3 +1,4 @@
+//go:build getall
 // +build getall
 
 /*
@@ -16,7 +17,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package internal
+package pkg
 
 import "strings"
