+++ cmd/version.go
@@ -19,7 +19,7 @@ package cmd
 import (
 	"text/template"
 
-	"github.com/corneliusweig/ketall/internal/version"
+	"github.com/corneliusweig/ketall/pkg/version"
 	"github.com/spf13/cobra"
 	"k8s.io/klog/v2"
 )
