+++ pkg/options/options.go
@@ -20,7 +20,7 @@ import (
 	"bytes"
 	"os"
 
-	"github.com/corneliusweig/ketall/internal/printer"
+	"github.com/corneliusweig/ketall/pkg/printer"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 	"k8s.io/cli-runtime/pkg/printers"
 	"k8s.io/klog/v2"
