+++ pkg/printer/adapter.go
