+++ pkg/printer/adapter_test.go
